refactor: move deletemutation to module
