+++ src/Apps/CollectorProfile/Routes/Saves2/Components/Actions/DeleteSavesModal.tsx
@@ -1,9 +1,7 @@
 import { Button, Flex, ModalDialog, Text, useToasts } from "@artsy/palette"
+import { useDeleteCollection } from "Apps/CollectorProfile/Routes/Saves2/Components/Actions/Mutations/useDeleteCollection"
 import { useTranslation } from "react-i18next"
-import { ConnectionHandler, graphql } from "react-relay"
 import { useRouter } from "System/Router/useRouter"
-import { useMutation } from "Utils/Hooks/useMutation"
-import { DeleteSavesModalMutation } from "__generated__/DeleteSavesModalMutation.graphql"
 import { SavesArtworks_collection$data } from "__generated__/SavesArtworks_collection.graphql"
 
 interface Props {
@@ -11,44 +9,11 @@ interface Props {
   onClose: () => void
 }
 
-const deleteCollectionUpdater = (store, collectionID) => {
-  const root = store.getRoot()
-  const me = root.getLinkedRecord("me")
-
-  const otherSavesConnection = ConnectionHandler.getConnection(
-    me,
-    "CollectorProfileSaves2Route_otherSaves"
-  )
-
-  if (!otherSavesConnection) {
-    return
-  }
-
-  ConnectionHandler.deleteNode(otherSavesConnection, collectionID)
-}
-
 export const DeleteSavesModal: React.FC<Props> = ({ collection, onClose }) => {
   const { t } = useTranslation()
   const { router } = useRouter()
 
-  const { submitMutation } = useMutation<DeleteSavesModalMutation>({
-    mutation: graphql`
-      mutation DeleteSavesModalMutation($input: deleteCollectionInput!) {
-        deleteCollection(input: $input) {
-          responseOrError {
-            __typename # DeleteCollectionSuccess or DeleteCollectionFailure
-            ... on DeleteCollectionFailure {
-              mutationError {
-                message
-                statusCode
-              }
-            }
-          }
-        }
-      }
-    `,
-    updater: store => deleteCollectionUpdater(store, collection.internalID),
-  })
+  const { submitMutation } = useDeleteCollection()
 
   const { sendToast } = useToasts()
